# specs/003-line-login-auth/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: LINE Login 會員認證系統 API
  description: |
    YouTube Loop Player 的 LINE Login 會員認證與資料管理 API

    ## 認證機制
    - 使用 HTTP-only cookies 儲存 access token
    - 每個請求自動攜帶 cookie,後端驗證 token 有效性
    - Token 過期時返回 401,前端應引導使用者重新登入

    ## 錯誤處理
    所有錯誤回應遵循統一格式:
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "使用者友善的錯誤訊息"
      }
    }
    ```

  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8080/api
    description: 開發環境
  - url: https://yourdomain.com/api
    description: 生產環境
...
paths:
...
  /playlists:
    get:
      tags:
        - Playlists
      summary: 取得會員的所有播放清單
      description: |
        取得當前會員建立的所有播放清單
        - 不包含播放清單項目 (僅基本資訊)
        - 按建立時間倒序排列
      operationId: getUserPlaylists
      security:
        - cookieAuth: []
      responses:
        '200':
          description: 成功取得播放清單
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Playlist'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Playlists
      summary: 建立新播放清單
      description: |
        建立一個新的播放清單
        - 名稱不可重複 (同一會員)
      operationId: createPlaylist
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  description: 播放清單名稱
                  example: 我的最愛
                  minLength: 1
                  maxLength: 255
                description:
                  type: string
                  description: 播放清單描述
                  example: 收藏的經典歌曲
                  nullable: true
                isPublic:
                  type: boolean
                  description: 是否公開
                  default: false
      responses:
        '201':
          description: 播放清單已建立
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Playlist'
        '400':
          description: 名稱已存在或無效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: PLAYLIST_NAME_EXISTS
                  message: 已存在相同名稱的播放清單
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: HTTP-only cookie 儲存的 access token
  schemas:
...
    Playlist:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 我的最愛
        description:
          type: string
          nullable: true
          example: 收藏的經典歌曲
        isPublic:
          type: boolean
          example: false
        itemCount:
          type: integer
          example: 10
          description: 播放清單中的影片數量
        createdAt:
          type: string
          format: date-time
          example: 2025-11-01T10:00:00Z
        updatedAt:
          type: string
          format: date-time
          nullable: true
          example: 2025-11-01T12:00:00Z
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: ERROR_CODE
            message:
              type: string
              example: 錯誤訊息描述
  responses:
    Unauthorized:
      description: 未授權 - Token 無效或已過期
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: 請先登入
...